2010-06-04  Ramon Novoa  <rnovoa@artica.es>
	* pandora_server.spec: Fixed the %postun target.




git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@2859 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_server/pandora_server.spec
@@ -119,6 +119,12 @@ chkconfig --del pandora_server
 chkconfig --del tentacle_serverd
 
 %postun
+
+# Upgrading
+if [ "$1" = "1" ]; then
+        exit 0
+fi
+
 rm -Rf /etc/init.d/tentacle_serverd
 rm -Rf /etc/init.d/pandora_server
 rm -Rf %{prefix}pandora_server
